Resolve MUWM-4231
