+++ myuw/models/__init__.py
@@ -300,9 +300,8 @@ class UserCourseDisplay(models.Model):
 
     @classmethod
     def delete_section_display(cls, user, section_label):
-        obj = UserCourseDisplay.objects.get(
-            user=user, section_label=section_label)
-        obj.delete()
+        UserCourseDisplay.objects.filter(
+            user=user, section_label=section_label).delete()
 
     @classmethod
     def get_section_display(cls, user, section_label):
