+++ myuw/test/dao/test_user_course_display.py
@@ -9,7 +9,7 @@ from myuw.dao.registration import get_schedule_by_term
 from myuw.dao.term import get_current_quarter
 from myuw.dao.user import get_user_model
 from myuw.dao.user_course_display import set_course_display_pref,\
-    _get_next_color, set_pin_on_teaching_page, _save_section_color
+    _get_next_color, _save_section_color
 from myuw.test import get_request_with_user, get_request_with_date
 
 
@@ -140,13 +140,12 @@ class TestUserCourseDisplayDao(TransactionTestCase):
         records = UserCourseDisplay.objects.all()
         self.assertEqual(len(records), 7)
 
-        #  test drop sections
-        schedule.sections.remove(sections[6])
-        schedule.sections.remove(sections[0])
-        schedule.sections.remove(sections[0])
-        set_course_display_pref(req, schedule)
+        # test delete a record
+        UserCourseDisplay.delete_section_display(
+            get_user_model(req),
+            sections[6].section_label())
         records = UserCourseDisplay.objects.all()
-        self.assertEqual(len(records), 4)
+        self.assertEqual(len(records), 6)
 
         # add them back
         schedule = get_instructor_schedule_by_term(req, term)
@@ -162,45 +161,3 @@ class TestUserCourseDisplayDao(TransactionTestCase):
         self.assertEqual(sections[6].color_id, '4a')
         records = UserCourseDisplay.objects.all()
         self.assertEqual(len(records), 7)
-
-    def test_set_pin_on_teaching_page(self):
-        req = get_request_with_user("billsea")
-        term = get_current_quarter(req)
-        schedule = get_instructor_schedule_by_term(req, term)
-        set_course_display_pref(req, schedule)
-        records = UserCourseDisplay.objects.all()
-        self.assertFalse(records[2].pin_on_teaching_page)
-        self.assertFalse(records[3].pin_on_teaching_page)
-
-        section_label = schedule.sections[2].section_label()
-        # test pin mini card
-        self.assertTrue(set_pin_on_teaching_page(req,
-                                                 section_label))
-        records = UserCourseDisplay.objects.all()
-        self.assertTrue(records[2].pin_on_teaching_page)
-
-        # close pin mini card
-        self.assertTrue(set_pin_on_teaching_page(req,
-                                                 section_label,
-                                                 pin=False))
-        records = UserCourseDisplay.objects.all()
-        self.assertFalse(records[2].pin_on_teaching_page)
-
-        # test if not in DB
-        records[2].delete()
-        with self.assertRaises(UserCourseDisplay.DoesNotExist):
-            set_pin_on_teaching_page(req, section_label, pin=True)
-
-        # not pin primary section
-        section_label = '2013,spring,PHYS,122/A'
-        self.assertFalse(set_pin_on_teaching_page(req, section_label))
-
-        # test InvalidSectionID
-        section_label = '2013,spring,PHYS,122/'
-        with self.assertRaises(InvalidSectionID):
-            set_pin_on_teaching_page(req, section_label)
-
-        # test NotSectionInstructorException
-        section_label = '2013,spring,PHYS,121/AC'
-        with self.assertRaises(NotSectionInstructorException):
-            set_pin_on_teaching_page(req, section_label)
